Start working out databases for project
I have created a way to generate the project databases.  These are
generated using SQLite3.  The SQL should be relatively standard, so
could be used for other DBMSs.  See setupexample.rb.  To generate the
databases, run

	ruby stepexample.rb

Otherwise, I updated the .gitignore to get rid of things that are not
relevent.